requirements.txt
# TwinBuddy Hackathon MVP — FastAPI 后端依赖
# Python >= 3.10

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.2
httpx==0.27.2
langgraph==0.2.56
pyyaml>=6.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
